more tab handling specs
