+++ extension/test-utils/mockActiveTab.ts
@@ -1,4 +1,5 @@
 import { vi } from 'vitest'
+import { createMockTab } from './createMockTab'
 
 type ResultFn = (tabs: chrome.tabs.Tab[]) => void
 
@@ -7,20 +8,7 @@ export const mockActiveTab = (tab: Partial<chrome.tabs.Tab> = {}) => {
 
   mockQuery.mockImplementation(
     (_: chrome.tabs.QueryInfo, callback?: ResultFn) => {
-      const activeTab = {
-        active: true,
-        autoDiscardable: true,
-        discarded: false,
-        groupId: 0,
-        highlighted: false,
-        incognito: false,
-        index: 0,
-        pinned: false,
-        windowId: 0,
-        selected: true,
-
-        ...tab,
-      } satisfies chrome.tabs.Tab
+      const activeTab = createMockTab(tab)
 
       if (typeof callback === 'function') {
         callback([activeTab])
